.idea/requirements.txt
flask==3.0.3
pytest==8.3.2
bandit==1.7.9
requests==2.32.3
python-dotenv==1.0.1